[feature]附件上传-》视频浏览，删除、浏览按钮控制

+++ app/components/form/attachment-control/attachment-list/attachment-list.js
@@ -116,13 +116,12 @@ let css=`
     .closeImg:hover{
         background-color:darkred;
     }
-
 `;
 
 export const attachmentListConfig = {
     template: template,
     data:{
-        css:css.replace(/(\n)/g, '')
+        css:css.replace(/(\n)/g, ''),
     },
     actions:{
         deleteItem:function (item) {
@@ -144,9 +143,30 @@ export const attachmentListConfig = {
             Storage.setItem(deletedFiles,'deletedItem-'+this.data.id,Storage.SECTION.FORM);
         },
         viewItem:function (item) {
-            let ele = $('<img>');
-            ele.attr('src','/download_attachment/?file_id='+item.file_id+'&download=0&dinput_type='+this.data.dinput_type);
+            if(item.viewable == 'no'){
+                return;
+            }
+            let url = '/download_attachment/?file_id='+item.file_id+'&download=0&dinput_type='+this.data.dinput_type;
+            let ele;
+            if(item.viewable == 'viewable'){
+                ele = $('<img src="'+url+'">');
+            } else if(item.viewable == 'playable'){
+                ele = $('<video width="400" controls><source src="'+url+'" type="video/mp4">您的浏览器不支持HTML5</video>');
+            }
             this.el.find('.preview-anchor').empty().append(ele);
+        },
+        isViewable:function (filename) {
+            let exp = filename.split('.').pop();
+            let viewableFileType = ['jpg','jpeg','bmp', 'png', 'tiff', 'gif', 'exif', 'svg', 'pcd', 'dxf', 'ufo'];
+            for(let type of viewableFileType){
+                if(exp.toLowerCase().indexOf(type)!=-1){
+                    return 'viewable';
+                }
+            }
+            if(exp.toLowerCase().indexOf('mp4')!=-1){
+                return 'playable';
+            }
+            return 'no';
         }
     },
     afterRender:function () {
@@ -163,8 +183,7 @@ export const attachmentListConfig = {
                     row['dinput_type']=this.data.dinput_type;
                     row['callback'] = this.actions.deleteItem;
                     let ele = $('<tr></tr>');
-                    let fileTd = $('<td></td>');
-                    fileTd.text(row.file_name);
+                    let fileTd = $('<td>'+row.file_name+'</td>');
                     ele.append(fileTd);
                     let controlersTd = $('<td></td>');
                     let downloadCon = $('<a>下载</a>');
@@ -175,14 +194,19 @@ export const attachmentListConfig = {
                     viewCon.on('click',function () {
                         t.actions.viewItem(row);
                     });
-                    let deleteCon = $('<a>删除</a>');
-                    controlersTd.append(deleteCon);
-                    deleteCon.on('click',function () {
-                        t.actions.deleteItem(row);
-                        ele.remove();
-                    })
+                    if(this.data.isView == 0){
+                        let deleteCon = $('<a>删除</a>');
+                        controlersTd.append(deleteCon);
+                        deleteCon.on('click',function () {
+                            t.actions.deleteItem(row);
+                            ele.remove();
+                        })
+                    }
                     ele.append(controlersTd);
-                    row['element']=ele;
+                    row['viewable']= this.actions.isViewable(row.file_name);
+                    if(row['viewable']=='no'){
+                        viewCon.css('color','#a9a9a9');
+                    }
                     this.el.find('.attachment-list-anchor').append(ele);
                 }
             }
